User creation, Authentication
Frontend
implemented state management, added the ability to login and register, protected certain routes so only logged in users can view them.

Backend-
dependencies added:
uuid for unique id's for users, bcrypt to hash passwords for safe storage in the DB, deep-email-validator to validate emails before a user is created, jsonwebtoken, this is how users are going to authenticate to protected routes on the backend api.